+++ src/icons.mli
@@ -1,1 +1,4 @@
+
 val drag_icon : GdkPixbuf.pixbuf
+
+val doculib_icon : GdkPixbuf.pixbuf